Moved updatePayments to PaymentsProcessor
